requirements.txt
discord.py==2.3.2
httpx==0.27.0
starlette==0.37.2
uvicorn[standard]==0.29.0
